Allow different module loading strategies trough an IModuleNamesProvider
- Introduce IApplicationContext replacing IHostingEnvironment.GetApplication()
- Introduce IModuleNamesProvider with a default implementation (AssemblyAttributeModuleNamesProvider) using ModuleNameAttribute
- Replace all occurrences of IHostingEnvironment.GetApplication()/GetModule() with IApplicationContext
