+++ src/OrchardCore/OrchardCore.DisplayManagement.Liquid/LiquidViewOptionsSetup.cs
@@ -1,20 +1,24 @@
 using System;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Options;
+using OrchardCore.Modules;
 
 namespace OrchardCore.DisplayManagement.Liquid
 {
     public class LiquidViewOptionsSetup : IConfigureOptions<LiquidViewOptions>
     {
         private readonly IHostingEnvironment _hostingEnvironment;
+        private readonly IApplicationContext _applicationContext;
 
         /// <summary>
         /// Initializes a new instance of <see cref="LiquidViewOptions"/>.
         /// </summary>
         /// <param name="hostingEnvironment"><see cref="IHostingEnvironment"/> for the application.</param>
-        public LiquidViewOptionsSetup(IHostingEnvironment hostingEnvironment)
+        /// <param name="applicationContext"><see cref="IApplicationContext"/> for the application.</param>
+        public LiquidViewOptionsSetup(IHostingEnvironment hostingEnvironment, IApplicationContext applicationContext)
         {
             _hostingEnvironment = hostingEnvironment ?? throw new ArgumentNullException(nameof(hostingEnvironment));
+            _applicationContext = applicationContext ?? throw new ArgumentNullException(nameof(applicationContext));
         }
 
         public void Configure(LiquidViewOptions options)
@@ -25,7 +29,7 @@ namespace OrchardCore.DisplayManagement.Liquid
 
                 if (_hostingEnvironment.IsDevelopment())
                 {
-                    options.FileProviders.Insert(0, new ModuleProjectLiquidFileProvider(_hostingEnvironment));
+                    options.FileProviders.Insert(0, new ModuleProjectLiquidFileProvider(_applicationContext));
                 }
             }
         }
