+++ src/OrchardCore/OrchardCore.DisplayManagement.Liquid/ModuleProjectLiquidFileProvider.cs
@@ -17,9 +17,9 @@ namespace OrchardCore.DisplayManagement.Liquid
     public class ModuleProjectLiquidFileProvider : IFileProvider
     {
         private static Dictionary<string, string> _paths;
-        private static object _synLock = new object();
+        private static readonly object _synLock = new object();
 
-        public ModuleProjectLiquidFileProvider(IHostingEnvironment environment)
+        public ModuleProjectLiquidFileProvider(IApplicationContext applicationContext)
         {
             if (_paths != null)
             {
@@ -33,12 +33,10 @@ namespace OrchardCore.DisplayManagement.Liquid
                     if (_paths == null)
                     {
                         var assets = new List<Asset>();
-                        var application = environment.GetApplication();
+                        var application = applicationContext.Application;
 
-                        foreach (var name in application.ModuleNames)
+                        foreach (var module in application.Modules)
                         {
-                            var module = environment.GetModule(name);
-
                             if (module.Assembly == null || Path.GetDirectoryName(module.Assembly.Location)
                                 != Path.GetDirectoryName(application.Assembly.Location))
                             {
