+++ src/OrchardCore/OrchardCore.Modules.Abstractions/AssemblyAttributeModuleNamesProvider.cs
@@ -1,0 +1,24 @@
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+using Microsoft.AspNetCore.Hosting;
+using OrchardCore.Modules.Manifest;
+
+namespace OrchardCore.Modules
+{
+    public class AssemblyAttributeModuleNamesProvider : IModuleNamesProvider
+    {
+        private readonly List<string> _moduleNames;
+
+        public AssemblyAttributeModuleNamesProvider(IHostingEnvironment hostingEnvironment)
+        {
+            var assembly = Assembly.Load(new AssemblyName(hostingEnvironment.ApplicationName));
+            _moduleNames = assembly.GetCustomAttributes<ModuleNameAttribute>().Select(m => m.Name).ToList();
+        }
+
+        public IEnumerable<string> GetModuleNames()
+        {
+            return _moduleNames;
+        }
+    }
+}
