+++ src/OrchardCore/OrchardCore.Modules.Abstractions/Asset.cs
@@ -1,0 +1,25 @@
+namespace OrchardCore.Modules
+{
+    public class Asset
+    {
+        public Asset(string asset)
+        {
+            asset = asset.Replace('\\', '/');
+            var index = asset.IndexOf('|');
+
+            if (index == -1)
+            {
+                ModuleAssetPath = string.Empty;
+                ProjectAssetPath = string.Empty;
+            }
+            else
+            {
+                ModuleAssetPath = asset.Substring(0, index);
+                ProjectAssetPath = asset.Substring(index + 1);
+            }
+        }
+
+        public string ModuleAssetPath { get; }
+        public string ProjectAssetPath { get; }
+    }
+}
