+++ src/OrchardCore/OrchardCore.Modules.Abstractions/IModuleNamesProvider.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+
+namespace OrchardCore.Modules
+{
+    public interface IModuleNamesProvider
+    {
+        IEnumerable<string> GetModuleNames();
+    }
+}
