+++ src/OrchardCore/OrchardCore.Modules/Extensions/ServiceCollectionExtensions.cs
@@ -87,6 +87,9 @@ namespace Microsoft.Extensions.DependencyInjection
 
         private static void AddExtensionServices(OrchardCoreBuilder builder)
         {
+            builder.ApplicationServices.AddSingleton<IModuleNamesProvider, AssemblyAttributeModuleNamesProvider>();
+            builder.ApplicationServices.AddSingleton<IApplicationContext, ModularApplicationContext>();
+            
             builder.ApplicationServices.AddExtensionManagerHost();
 
             builder.ConfigureServices(services =>
@@ -103,18 +106,19 @@ namespace Microsoft.Extensions.DependencyInjection
             builder.Configure((app, routes, serviceProvider) =>
             {
                 var env = serviceProvider.GetRequiredService<IHostingEnvironment>();
+                var appContext = serviceProvider.GetRequiredService<IApplicationContext>();
 
                 IFileProvider fileProvider;
                 if (env.IsDevelopment())
                 {
                     var fileProviders = new List<IFileProvider>();
-                    fileProviders.Add(new ModuleProjectStaticFileProvider(env));
-                    fileProviders.Add(new ModuleEmbeddedStaticFileProvider(env));
+                    fileProviders.Add(new ModuleProjectStaticFileProvider(appContext));
+                    fileProviders.Add(new ModuleEmbeddedStaticFileProvider(appContext));
                     fileProvider = new CompositeFileProvider(fileProviders);
                 }
                 else
                 {
-                    fileProvider = new ModuleEmbeddedStaticFileProvider(env);
+                    fileProvider = new ModuleEmbeddedStaticFileProvider(appContext);
                 }
 
                 app.UseStaticFiles(new StaticFileOptions
