+++ test/OrchardCore.Tests/Extensions/ExtensionManagerTests.cs
@@ -1,9 +1,11 @@
+using System.Collections.Generic;
 using System.Linq;
 using Microsoft.AspNetCore.Hosting;
 using OrchardCore.DisplayManagement.Events;
 using OrchardCore.DisplayManagement.Extensions;
 using OrchardCore.Environment.Extensions;
 using OrchardCore.Environment.Extensions.Features;
+using OrchardCore.Modules;
 using OrchardCore.Tests.Stubs;
 using Xunit;
 
@@ -13,6 +15,12 @@ namespace OrchardCore.Tests.Extensions
     {
         private static IHostingEnvironment HostingEnvironment
             = new StubHostingEnvironment();
+
+        private static IApplicationContext ApplicationContext
+            = new ModularApplicationContext(HostingEnvironment, new List<IModuleNamesProvider>()
+            {
+                new AssemblyAttributeModuleNamesProvider(HostingEnvironment)
+            });
 
         private static IFeaturesProvider ModuleFeatureProvider =
             new FeaturesProvider(new[] { new ThemeFeatureBuilderEvents() }, new NullLogger<FeaturesProvider>());
@@ -28,6 +36,7 @@ namespace OrchardCore.Tests.Extensions
         {
             ModuleScopedExtensionManager = new ExtensionManager(
                 HostingEnvironment,
+                ApplicationContext,
                 new[] { new ExtensionDependencyStrategy() },
                 new[] { new ExtensionPriorityStrategy() },
                 new TypeFeatureProvider(),
@@ -37,6 +46,7 @@ namespace OrchardCore.Tests.Extensions
 
             ThemeScopedExtensionManager = new ExtensionManager(
                 HostingEnvironment,
+                ApplicationContext,
                 new[] { new ExtensionDependencyStrategy() },
                 new[] { new ExtensionPriorityStrategy() },
                 new TypeFeatureProvider(),
@@ -46,6 +56,7 @@ namespace OrchardCore.Tests.Extensions
 
             ModuleThemeScopedExtensionManager = new ExtensionManager(
                 HostingEnvironment,
+                ApplicationContext,
                 new IExtensionDependencyStrategy[] { new ExtensionDependencyStrategy(), new ThemeExtensionDependencyStrategy() },
                 new[] { new ExtensionPriorityStrategy() },
                 new TypeFeatureProvider(),
